Finished the Delelte Bet Modal
